+++ src/lang/prog/script/embed/mod.rs
@@ -1,0 +1,19 @@
+// devela/src/lang/prog/script/embed/mod.rs
+//
+#![doc = crate::_tags!(wip)]
+#![doc = crate::_DOC_LANG_PROG_SCRIPT_EMBED!()] // public
+#![doc = crate::_doc!(modules: crate::lang::prog::script; embed)]
+#![doc = crate::_doc!(flat:"lang")]
+#![doc = crate::_doc!(hr)]
+//
+
+// mod script; // WIP script!
+
+crate::structural_mods! { // _mods
+    _mods {
+        // #[doc(inline)]
+        // pub use super::{
+        //     script::script,
+        // };
+    }
+}