+++ src/lang/prog/script/form/shell/word/_test.rs
@@ -1,4 +1,4 @@
-// devela/src/lang/prog/script/shell/word/_test.rs
+// devela/src/lang/prog/script/form/shell/word/_test.rs
 
 use crate::{ShellLex, ShellQuote, ShellWordError};