+++ src/lang/prog/script/form/shell/word/quote.rs
@@ -1,4 +1,4 @@
-// devela/src/lang/prog/script/shell/word/quote.rs
+// devela/src/lang/prog/script/form/shell/word/quote.rs
 //
 //! Shell word parsing and quoting.
 //
@@ -7,7 +7,10 @@ use crate::{ShellWordError, is, whilst, write_at};
 
 #[doc = crate::_tags!(lang text)]
 /// Quotes shell words for command strings.
-#[doc = crate::_doc_meta!{location("lang/prog/script/shell")}]
+#[doc = crate::_doc_meta!{
+    location("lang/prog/script/form/shell"),
+    test_size_of(ShellQuote = 1|8),
+}]
 #[must_use]
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Hash)]
 pub struct ShellQuote {